:recycle: rewrite `process_hints()`
Rewrote `process_hints()` to use imperative style instead of
functional. I think, in this style, I can incorporate handling the B
hints better. Only time will tell if that gut feeling was correct.

If the closure to `.fold()` was simpler, I would prefer the functional
style. But that closure kept getting larger, so it's easier to read as
the body of a `for-loop`.