:sparkles: merge in acceptable words
Up to this point, the program was using the dictionary of all possible
solutions. This is unfair because all of its guesses will be in this
set. A person, on the other hand, has tens of thousands of words.

The offcial Wordle game has a dictionary of words it'll accept as a
guess -- even if it isn't in the set of possible answers. This commit
pulls in this set of words so the program has the same number of
inputs a person has. (12,000+ words!)