:recycle: "dedup" after "sorting"
When picking "best" words for the guess, words with double letters
were being filtered. The theory that words with 5 unique letters will
give us more clues. The problem was the letters were being dedup-ed
before sorting them so the program started the game with PAPAS, which
is a *terrible* first guess.

This commit sorts the letters and then removes duplicates.