:sparkles: allow double letters in small vocabularies
I had a wordle puzzle with a double letter. Wordlebot was down to one
letter to find and 5 choices. Because of the "no double letter" filter,
the correct answer was chosen last.

This commit allows words with double letters to be considered if the
remaining vocabulary has less than 100 words.